chore(routines): split command_tests.rs binary-resolution tests out
src/routines/command_tests.rs had crept to 713 lines, past the repo's
700-line pre-push gate. Moves the 18 tmux/agent-binary-resolution
tests into a new command_bin_resolution_tests.rs sibling, matching the
existing one-helper-copy-per-test-file convention. No behavior change;
same 886 tests pass, command.rs coverage unchanged at 100%.
